Rerecord round join to include everything from the start
I need this dump to look at the snapshots. So to undelta the
correct snapshot I need every single snap sent from the very beginning.

The demo, log and pcap are updated. I reused the old images which are now
technically wrong. But it is just tees spawning without inputs
so the pictures should still be quite accurate actually.